metaServer/requirements.txt
mcp>=0.1.0
python-docx>=1.0.0